[cm] Add filterbanks file, remove 1d channel summation
